// src/components/global/Ambilight/ambilight.interfaces.ts
export interface AmbilightProps {
    imagesUrls: string[]
    colors: number
    aspectRatio: number
    currentIndex: number
    className?: string,
    fadeDuration?: number
}

// export interface AmbilightProps {
//     className?: string
//     imageUrl: string
//     width: number
//     height: number
// }